fix(core): resolve Random(BigDecimal) crash and LIP resolution failures

+++ src/main/groovy/com/endlesstransit/core/Game.groovy
@@ -753,7 +753,7 @@ class Game {
         lines << " " + Terminal.dim("[QUANTUM_SPECTROGRAM]")
         
         // Simple ASCII spectrogram based on inventory
-        Random r = new Random(System.currentTimeMillis() / 1000)
+        Random r = new Random((System.currentTimeMillis() / 1000) as long)
         for (int i = 0; i < 5; i++) {
             int h = r.nextInt(width / 4) + 1
             lines << " " + Terminal.colorize("█" * h, Terminal.CYAN)
